groq /requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket support
websockets==12.0

# HTTP client for API calls  
aiohttp==3.9.1
httpx==0.25.2

# Groq AI client
groq==0.4.1

# Environment variables
python-dotenv==1.0.0

# Audio processing (optional, for future enhancements)
pydub==0.25.1
numpy==1.24.3

# Async utilities
asyncio-mqtt==0.13.0

# Development and debugging
python-multipart==0.0.6

# Pipecat dependencies (if using the full framework)
# Note: Some of these might be optional depending on your Pipecat setup
# pipecat-ai==0.0.40
# daily-python==0.10.1

# Additional utilities for production
structlog==23.2.0
colorama==0.4.6

# Optional: For enhanced audio processing
# librosa==0.10.1
# soundfile==0.12.1